Restructure of Draggable
