+++ src/Draggable/index.js
@@ -1,0 +1,3 @@
+import Draggable from './Draggable';
+
+export default Draggable;
