+++ src/droppable.js
@@ -1,4 +1,4 @@
-import Draggable from './draggable';
+import Draggable from './Draggable';
 import {closest} from './utils';
 
 import {
